P3-T3

+++ argc_argv/3-mul.c
@@ -10,7 +10,7 @@
 
 int main(int argc, char *argv[])
 {
-	int num1, num2, result;
+	long int num1, num2, result;
 
 	if (argc != 3)
 	{
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 
 	result = num1 * num2;
 
-	printf("%d\n", result);
+	printf("%ld\n", result);
 
 	return (0);
 }
